:wrench: chore: start consolidating provider enum
